move connection tests
