+++ tests/connection/test_base.py
